+++ packages/content/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.5
+
+- Add option to prerender markdown format content items as html by appending the query param ?render-html
+- Fix major admin UI bug introducted in 0.2.2 that caused posting from the create-content-item-modal to fail because it was sending a redundant field that caused validation to reject
+
 ## 0.2.4
 
 - Updated medusa packages to 2.17.2